pyproject.toml
[tool.poetry]
name = "rag-quiz"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
pandas = "^2.2.0"
numpy = "^1.26.3"
scipy = "^1.12.0"
jupyterlab = "^4.0.12"
matplotlib = "^3.8.2"
seaborn = "^0.13.2"
graphviz = "^0.20.1"
ipykernel = "^6.29.4"
logging = "^0.4.9.6"
papermill = "^2.6.0"
pre-commit = "^3.7.1"
nbstripout = "^0.7.1"
openai = "^1.59.6"
python-dotenv = "^1.0.1"
sqlalchemy = "^2.0.37"
psycopg2-binary = "^2.9.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

 
[tool.ruff]
line-length = 120
 
 
[tool.ruff.lint]
select = [
    "E", # pep8
    "F", # flake8 (base)
    "B", # flake8-bugbear
    "SIM", # flake8-simplify -> better ways to code thing
    "I", # flake8-imports: check order and formatting according to pep8
    #"D", # flake8-docstrings
    "T", # flake8-print - check for print statements left in code
    "N", # pep8-naming,
    "ANN", # flake8-annotations
    "PD", # pandas-vet
    "C", #flakeu-comprehensions
    "PT", # flake8-pytest-style
    "RET", # flake8-return
    "TD", # flake8-todos
    "TCH", # flake8-type-checking
    "FBT", # flake8-boolean-trap
]  
ignore = ["D104"] # ignore docstring in __init__.py
 
[tool.ruff.lint.pydocstyle]
convention = "google"
 
[tool.ruff.format]
line-ending = "lf"
skip-magic-trailing-comma = true
 
[tool.ruff.lint.isort]
split-on-trailing-comma = false
no-sections = true
